refactor to simplify main, lint.
