+++ packages/cloudbuster/src/breakglass/types.ts
@@ -1,0 +1,7 @@
+export interface BreakglassUser {
+	accountName: string;
+	user: string;
+	userUrl: string;
+	mfaActive: boolean;
+	hasUsernameTag: boolean;
+}
