+++ packages/cloudbuster/src/index.ts
@@ -1,196 +1,15 @@
 import { SNSClient } from '@aws-sdk/client-sns';
-import {
-	Anghammarad,
-	AnghammaradNotification,
-	RequestedChannel,
-	Target,
-} from '@guardian/anghammarad';
+import { Anghammarad } from '@guardian/anghammarad';
 import { awsClientConfig } from 'common/aws.js';
 import { logger } from 'common/logs.js';
-import type {
-	cloudbuster_fsbp_vulnerabilities,
-	PrismaClient,
-} from 'common/prisma-client/client.js';
-import {
-	getAwsAccounts,
-	getFsbpFindings,
-	getIamCredentialReports,
-	getIamUsers,
-} from 'common/src/database-queries.js';
 import { getPrismaClient } from 'common/src/prisma-client-setup.js';
-import type { SecurityHubSeverity } from 'common/src/types.js';
-import { BreakglassUser, createBreakglassUserReport } from './breakglass.js';
-import type { Config } from './config.js';
+import { sendBreakglassUserAlerts } from './breakglass/index.js';
 import { getConfig } from './config.js';
-import { createDigestsFromFindings, sendDigest } from './digests.js';
-import { findingsToGuardianFormat } from './findings.js';
-
-async function createFsbpTableAndAlerts(
-	config: Config,
-	prisma: PrismaClient,
-	anghammaradClient: Anghammarad,
-) {
-	const severities: SecurityHubSeverity[] = ['CRITICAL', 'HIGH'];
-
-	// *** DATA GATHERING ***
-	logger.log({
-		message: `Level of severities that will be scanned: ${severities.join(', ')}`,
-	});
-
-	const dbResults = await getFsbpFindings(prisma, severities);
-
-	const tableContents: cloudbuster_fsbp_vulnerabilities[] = dbResults.flatMap(
-		findingsToGuardianFormat,
-	);
-
-	const controlIdArns = new Map<string, cloudbuster_fsbp_vulnerabilities>();
-	const duplicateControlIdArns: string[] = [];
-	tableContents.forEach((row) => {
-		const compositeKey = `${row.control_id}:${row.arn}:${row.aws_region}`;
-		if (controlIdArns.has(compositeKey)) {
-			if (!duplicateControlIdArns.includes(compositeKey)) {
-				duplicateControlIdArns.push(compositeKey);
-			}
-		} else {
-			controlIdArns.set(compositeKey, row);
-		}
-	});
-	const uniqueTableContents = Array.from(controlIdArns.values());
-
-	logger.log({
-		message: `${tableContents.length} high and critical FSBP findings detected`,
-	});
-
-	if (tableContents.length !== uniqueTableContents.length) {
-		logger.warn({
-			message: `${tableContents.length - uniqueTableContents.length} duplicate FSBP findings detected with control IDs and resource ARNs, including: ${duplicateControlIdArns.slice(0, 10).join(', ')}`,
-		});
-	}
-
-	await prisma.cloudbuster_fsbp_vulnerabilities.deleteMany();
-	await prisma.cloudbuster_fsbp_vulnerabilities.createMany({
-		data: uniqueTableContents,
-	});
-
-	const activeFindings: cloudbuster_fsbp_vulnerabilities[] =
-		uniqueTableContents.filter((row) => !row.suppressed);
-
-	logger.log({ message: `${activeFindings.length} active FSBP findings.` });
-
-	logger.log({
-		message: `Creating digests for 'CRITICAL' severity findings.`,
-	});
-
-	const digests = createDigestsFromFindings(
-		activeFindings,
-		'CRITICAL',
-		config.cutOffInDays,
-	);
-
-	const isTuesday = new Date().getDay() === 2;
-	if (isTuesday) {
-		logger.log({ message: "Creating digests for 'HIGH' severity findings." });
-		digests.push(
-			...createDigestsFromFindings(activeFindings, 'HIGH', config.cutOffInDays),
-		);
-	}
-
-	// *** NOTIFICATION SENDING ***
-	logger.log({ message: `Sending ${digests.length} digests.` });
-	await Promise.all(
-		digests.map(
-			async (digest) => await sendDigest(anghammaradClient, config, digest),
-		),
-	);
-}
-
-function formatUser(user: BreakglassUser): string {
-	const mfaString = user.mfaActive ? '' : 'MFA not active';
-	const tagString = user.hasUsernameTag ? '' : 'GoogleUsername tag not present';
-
-	const issues = [mfaString, tagString].filter(Boolean).join(', ');
-	return `[${user.user}](${user.userUrl}) - ${issues}.`;
-}
-
-function formatMessage(users: BreakglassUser[]): string {
-	return users.map(formatUser).join('\n');
-}
-
-async function breakglassUserReport(
-	config: Config,
-	prisma: PrismaClient,
-	anghammaradClient: Anghammarad,
-) {
-	const [credentialReports, awsAccounts, iamUsers] = await Promise.all([
-		getIamCredentialReports(prisma),
-		getAwsAccounts(prisma),
-		getIamUsers(prisma),
-	]);
-
-	const report = createBreakglassUserReport(
-		credentialReports,
-		awsAccounts,
-		iamUsers,
-	);
-
-	//TODO store user count as a cloudwatch metric
-
-	console.table(
-		report.map(({ user, mfaActive, hasUsernameTag }) => ({
-			user,
-			mfaActive,
-			hasUsernameTag,
-		})),
-	);
-
-	const usersPerAccount = awsAccounts
-		.map((account) => ({
-			acctId: account.id,
-			acctName: account.name,
-			users: report
-				.filter(
-					(user) => user.accountName === account.name && user.user != null,
-				)
-				.map((user) => user),
-		}))
-		.filter((account) => account.users.length > 0);
-
-	console.table(
-		usersPerAccount.map(({ acctName, users }) => ({
-			acctName,
-			numUsers: users.length,
-			users: users.map((u) => u.user).join(', '),
-		})),
-	);
-
-	const target: Target =
-		config.stage === 'PROD'
-			? { Stack: 'security' }
-			: { Stack: 'testing-alerts' };
-
-	usersPerAccount.forEach(async (account) => {
-		const notification: AnghammaradNotification = {
-			subject: `Breakglass User Report: ${account.acctName}`,
-			message: `${account.users.length} breakglass users are missing security configuration\n\n${formatMessage(account.users)}`,
-			actions: [
-				{
-					cta: 'View breakglass user report',
-					url: `https://metrics.gutools.co.uk/d/bdn97cui5rbi8f/var-account_name=${account.acctName}`,
-				},
-			],
-			target,
-			sender: `Cloudbuster ${config.stage}`,
-			channel: RequestedChannel.PreferHangouts,
-		};
-		await anghammaradClient.notify(notification);
-	});
-
-	return report;
-}
+import { createFsbpTableAndAlerts } from './fsbp/index.js';
 
 export async function main() {
 	logger.log({ message: 'Cloudbuster run starting.' });
-	// *** SETUP ***
+
 	const config = await getConfig();
 	const prisma = getPrismaClient(config);
 	const snsClient = new SNSClient(awsClientConfig(config.stage));
@@ -200,6 +19,7 @@ export async function main() {
 	);
 
 	await createFsbpTableAndAlerts(config, prisma, anghammaradClient);
-	await breakglassUserReport(config, prisma, anghammaradClient);
+	await sendBreakglassUserAlerts(config, prisma, anghammaradClient);
+
 	logger.log({ message: 'Cloudbuster run completed.' });
 }
